requirements.txt
# Core dependencies
llama-index>=0.9.0
faiss-cpu>=1.7.4  # Use faiss-gpu for CUDA support
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Machine Learning & Embeddings
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.22.0  # For optimized transformer operations

# Data Processing
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
pyarrow>=12.0.1  # For efficient DataFrame operations

# Utilities
rich>=13.4.2  # For enhanced terminal output
typing-extensions>=4.7.1
cachetools>=5.3.1
python-dateutil>=2.8.2

# Development dependencies
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.4.1
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.2
mkdocs-material>=9.1.21
mkdocstrings>=0.22.0